Adjust grid drawing logic
Originally, the individual grid items were given a defined width and height, and the overall size of the grid was calculated based on number of grid items in the grid. Now, the overall grid width is set up front, and the width and height of the individual grid items is calculated by this grid width and the number of grid items in the grid.
Essentially, this change allows the overall grid size to stay the same in all circumstances while the individual grid items can grow or shrink depending on how many grid items are in the grid.